travis: enable unconvert

+++ blockchain/chaingen/generator.go
@@ -427,7 +427,7 @@ func purchaseCommitmentScript(addr dcrutil.Address, amount, voteFeeLimit, revoca
 	copy(data[:], addr.ScriptAddress())
 	binary.LittleEndian.PutUint64(data[20:], uint64(amount))
 	data[27] |= 1 << 7
-	binary.LittleEndian.PutUint16(data[28:], uint16(limits))
+	binary.LittleEndian.PutUint16(data[28:], limits)
 	script, err := txscript.NewScriptBuilder().AddOp(txscript.OP_RETURN).
 		AddData(data[:]).Script()
 	if err != nil {
@@ -1066,7 +1066,7 @@ func (hp *hash256prng) uniformRandom(upperBound uint32) uint32 {
 	// (2^32 - (x*2)) % x == 2^32 % x when x <= 2^31
 	min := ((math.MaxUint32 - (upperBound * 2)) + 1) % upperBound
 	if upperBound > 0x80000000 {
-		min = uint32(1 + ^upperBound)
+		min = 1 + ^upperBound
 	}
 
 	r := hp.Hash256Rand()
