+++ mempool/mempool.go
@@ -1504,7 +1504,7 @@ func (mp *TxPool) RawMempoolVerbose(filterType *stake.TxType) map[string]*dcrjso
 			Size:             int32(tx.MsgTx().SerializeSize()),
 			Fee:              dcrutil.Amount(desc.Fee).ToCoin(),
 			Time:             desc.Added.Unix(),
-			Height:           int64(desc.Height),
+			Height:           desc.Height,
 			StartingPriority: desc.StartingPriority,
 			CurrentPriority:  currentPriority,
 			Depends:          make([]string, 0),
